replace the peer transfer example for something more practical
