+++ taskvine/src/examples/Makefile
@@ -4,7 +4,7 @@ include ../../../rules.mk
 LOCAL_LINKAGE+=${CCTOOLS_HOME}/taskvine/src/manager/libtaskvine.a ${CCTOOLS_HOME}/dttools/src/libdttools.a
 LOCAL_CCFLAGS+=-I ${CCTOOLS_HOME}/taskvine/src/manager
 
-PROGRAMS = vine_example_gutenberg vine_example_blast vine_example_mosaic vine_example_watch vine_example_minitask vine_example_poncho vine_example_peer_transfer
+PROGRAMS = vine_example_gutenberg vine_example_blast vine_example_mosaic vine_example_watch vine_example_minitask vine_example_poncho vine_example_peer_transfer vine_peer_slackware
 TARGETS = $(PROGRAMS)
 
 all: $(TARGETS)
